requirements.txt
apturl==0.5.2
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.6
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-rest-auth==2.1.4
Django==3.1.6
django-allauth==0.44.0
django-crispy-forms==1.11.1
django-extensions==3.1.1
django-jsonfield==1.4.1
djangorestframework==3.12.2
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
h11==0.12.0
httpcore==0.12.3
httplib2==0.14.0
httpx==0.17.0
idna==2.8
isort==5.8.0
jsonfield==3.1.0
keyring==18.0.1
keyrings.alt==3.4.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
mysqlclient==2.0.3
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pipreqs==0.4.10
praw==7.2.0
prawcore==2.0.0
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.7.0
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.7.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rfc3986==1.4.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sqlparse==0.4.1
systemd-python==234
tekore==3.6.1
toml==0.10.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
update-checker==0.18.0
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
websocket-client==0.58.0
wrapt==1.12.1
xkit==0.0.0
yarg==0.1.9
